+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Upcoming
 
+### Added
+
+- [[#244](https://github.com/rust-vmm/vmm-sys-util/pull/244)]: Add EventNotifier and EventConsumer as a generic event notification 
+
 ## v0.14.0
 
 ### Changed